fast-prof/requirements.txt
fastapi==0.111.0
minio==7.2.7
python-dotenv